maplebot/requirements.txt
# Pip requirements file
#

# WyzeSense2MQTT requirements
pyyaml
aiohttp
discord
discord-py-slash-command